+++ Tests/Integration/DSCResources/IisSite.config.ps1
@@ -2,18 +2,6 @@ Configuration IisSite_config
 {
     param
     (
-        [Parameter()]
-        [string[]]
-        $WebAppPool,
-
-        [Parameter(Mandatory = $true)]
-        [string[]]
-        $WebSiteName,
-
-        [Parameter(Mandatory = $true)]
-        [string]
-        $OsVersion,
-
         [Parameter(Mandatory = $true)]
         [string]
         $StigVersion,
@@ -28,7 +16,76 @@ Configuration IisSite_config
 
         [Parameter()]
         [string[]]
-        $Exception
+        $Exception,
+
+        [Parameter()]
+        [string[]]
+        $OrgSettings,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $BrowserVersion,
+
+        [Parameter()]
+        [AllowNull()]
+        [string[]]
+        $OfficeApp,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $ConfigPath,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $PropertiesPath,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $SqlVersion,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $SqlRole,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $ForestName,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $DomainName,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $OsVersion,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $OsRole,
+
+        [Parameter()]
+        [AllowNull()]
+        [string[]]
+        $WebAppPool,
+
+        [Parameter()]
+        [AllowNull()]
+        [string[]]
+        $WebSiteName,
+
+        [Parameter()]
+        [AllowNull()]
+        [string]
+        $LogPath
     )
 
     Import-DscResource -ModuleName PowerStig
@@ -39,22 +96,26 @@ Configuration IisSite_config
             {
                 $(if ($null -ne $WebAppPool)
                 {
-                   "WebAppPool = @($( ($WebAppPool | % {"'$_'"}) -join ',' ))`n"
+                   "WebAppPool = @($( ($WebAppPool | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
                 })
-                $( "WebSiteName = @($( ($WebSiteName | % {"'$_'"}) -join ',' ))`n" )
+                $( "WebSiteName = @($( ($WebSiteName | ForEach-Object {"'$PSItem'"}) -join ',' ))`n" )
                 OsVersion = '$OsVersion'
                 StigVersion = '$StigVersion'
+                $(if ($null -ne $OrgSettings)
+                {
+                    "Orgsettings = '$OrgSettings'"
+                })
                 $(if ($null -ne $Exception)
                 {
-                    "Exception = @{'$Exception'= @{'Value'='1234567'}}"
+                    "Exception = @{$( ($Exception | ForEach-Object {"'$PSItem'= @{'Value'='1234567'}"}) -join "`n" )}"
                 })
                 $(if ($null -ne $SkipRule)
                 {
-                    "SkipRule = @($( ($SkipRule | % {"'$_'"}) -join ',' ))`n"
+                    "SkipRule = @($( ($SkipRule | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
                 }
                 if ($null -ne $SkipRuleType)
                 {
-                    "SkipRuleType = @($( ($SkipRuleType | % {"'$_'"}) -join ',' ))`n"
+                    "SkipRuleType = @($( ($SkipRuleType | ForEach-Object {"'$PSItem'"}) -join ',' ))`n"
                 })
             }")
         )